+++ rekorder/lib/timestamp.py
@@ -4,6 +4,10 @@ import time
 
 class Timestamp:
 
+  @staticmethod
+  def playback_instance(*args, **kwargs):
+    return Timestamp(*args, **kwargs)
+
   def __init__(self, *args, **kwargs):
     super().__init__()
     self.time = kwargs.get('time', time.time())
@@ -11,3 +15,10 @@ class Timestamp:
 
   def __call__(self):
     return self
+
+  @staticmethod
+  def recordable_data(obj):
+    return {
+        'time': obj.time,
+        'localtime': obj.localtime
+    }